+++ .circleci/config.yml
@@ -11,5 +11,5 @@ workflows:
   maven_test:
     jobs:
       - maven/test:
-          test_results_path: /home/circleci/project/backend/target/surefire-reports
+          test_results_path: /home/circleci/project/app/target/surefire-reports
           executor: java19
