+++ app/src/main/java/ch/softwareplus/blueprints/config/AppConfig.java
@@ -1,0 +1,14 @@
+package ch.softwareplus.blueprints.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.hateoas.config.EnableHypermediaSupport;
+import org.springframework.hateoas.config.EnableHypermediaSupport.HypermediaType;
+
+/**
+ * Java config that contains the main application configuration.
+ */
+@Configuration
+@EnableHypermediaSupport(type = HypermediaType.HAL)
+public class AppConfig {
+    // intentionally empty
+}
