+++ app/src/main/java/ch/softwareplus/blueprints/config/SpringDocConfig.java
