+++ app/src/main/java/ch/softwareplus/blueprints/hero/HeroMapper.java
