+++ app/src/main/java/ch/softwareplus/blueprints/hero/HeroServiceImpl.java
