+++ app/src/main/java/ch/softwareplus/blueprints/hero/api/Hero.java
