+++ app/src/main/java/ch/softwareplus/blueprints/hero/api/HeroNotFoundException.java
