+++ app/src/main/java/ch/softwareplus/blueprints/hero/domain/HeroEntity.java
