+++ app/src/main/java/ch/softwareplus/blueprints/hero/rest/HeroController.java
