+++ app/src/main/java/ch/softwareplus/blueprints/hero/rest/HeroModelAssembler.java
