+++ app/src/main/java/ch/softwareplus/blueprints/hero/rest/HeroModelMapper.java
