+++ pom.xml
@@ -27,8 +27,6 @@
         <spring-restdocs-asciidoctor.version>3.0.0</spring-restdocs-asciidoctor.version>
         <spring-cloud.version>2022.0.1</spring-cloud.version>
 
-        <liquibase.version>4.18.0</liquibase.version>
-
         <!-- Plugins -->
         <maven-compiler-plugin.version>3.10.1</maven-compiler-plugin.version>
         <maven-surefire-plugin.version>2.22.2</maven-surefire-plugin.version>
@@ -45,7 +43,7 @@
 
 
     <modules>
-        <module>tour-of-heroes</module>
+        <module>app</module>
     </modules>
 
     <dependencyManagement>
