--- tour-of-heroes/src/main/resources/application.yml
@@ -1,61 +1,0 @@
-spring:
-  main:
-    banner-mode: "off"
-
-  jpa:
-    open-in-view: false
-    show-sql: false
-    hibernate:
-      ddl-auto: validate
-      naming.physical-strategy: org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-    properties:
-      hibernate:
-        format_sql: true
-
-  h2:
-    console:
-      enabled: true
-      path: /h2-console
-      settings.web-allow-others: true
-
-  datasource:
-    url: jdbc:h2:mem:heroes;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;
-    driverClassName: org.h2.Driver
-    username: sa
-    password:
-
-
-liquibase.change-log: classpath:db/changelog/db.changelog-master.yaml
-
-management:
-  endpoints:
-    web:
-      exposure:
-        include:
-          - health
-          - info
-          - prometheus
-
-server:
-  port: ${PORT:8080}
-  error:
-    include-stacktrace: never
-  forward-headers-strategy: native
-  servlet:
-    encoding:
-      force: true
-    context-path: /
-
-springdoc:
-  swagger-ui.path: /swagger-ui.html
-
-
-logging:
-  pattern:
-    console: "%d{HH:mm:ss.SSS} [%thread] %-5level %logger{35} - %msg %n"
-  level:
-    ROOT: INFO
-    # org.apache.http.wire: DEBUG
-    # org.hibernate.type.descriptor.sql.BasicBinder: TRACE
-    ch.softwareplus.blueprints: INFO
-    org.springframework.web: INFO
